backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pillow==10.1.0
face-recognition==1.3.0
qrcode==7.4.2
alembic==1.12.1
python-dotenv==1.0.0
# For PostgreSQL later
# psycopg2-binary==2.9.9